refactor: cleanup early implementation refactoring and second interface
- Removes second interface implementd for docker integrations
- Removes refactoring to core to call the new interface
