+++ common/protocol/src/types/interfaces/index.ts
@@ -1,5 +1,4 @@
 export * from "./cache.interface";
 export * from "./compression.interface";
 export * from "./runtime.interface";
-export * from "./grpc.runtime.interface";
 export * from "./storageProvider.interface";
