+++ integrations/docker/src/proto/runtime.proto
@@ -1,9 +1,7 @@
 syntax = "proto3";
 
-package kyve;
-
 import "google/protobuf/empty.proto";
-import "google/protobuf/any.proto";
+
 
 message DataItem {
   string key = 1;
