+++ integrations/tendermint-grpc/package.json
@@ -1,23 +1,13 @@
 {
-  "name": "@kyvejs/tendermint-grpc",
-  "version": "1.0.2",
+  "name": "@docker/tendermint",
+  "version": "1.0.0",
   "license": "MIT",
   "scripts": {
     "build": "rimraf dist && tsc",
-    "build:binaries": "yarn build && rimraf out && pkg --no-bytecode --public-packages '*' --output out/kyve package.json",
-    "start": "node ./src/runtime.js",
+    "start": "node ./dist/index.js",
     "format": "prettier --write ."
   },
-  "bin": "./src/runtime.js",
-  "pkg": {
-    "scripts": "./src/runtime.js",
-    "targets": [
-      "latest-linux-x64",
-      "latest-linux-arm64",
-      "latest-macos-x64"
-    ],
-    "outputPath": "out"
-  },
+  "bin": "./src/index.js",
   "prettier": {
     "singleQuote": true
   },
@@ -28,7 +18,6 @@
     "dotenv": "^16.3.1"
   },
   "devDependencies": {
-    "pkg": "^5.8.0",
     "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
     "typescript": "^4.7.4"
